platform_tests: Fixes and enhancements
- Added checks for crashes in all the tests.
- On failure, records all recent log msgs from all processes.
- Various fixes so that the tests work with new arch.

+++ platform_tests/testlib.py
@@ -3,14 +3,25 @@ import os
 import shutil
 import tempfile
 import pwd
+import time
 
 _NORMUSER = 'normuser'
+_RUNNER_LOG_FILE_FOR_ROOT = '/root/.jobber-log'
+_RUNNER_LOG_FILE_FOR_NORMUSER = '/home/{0}/.jobber-log'.\
+    format(_NORMUSER) 
 
 def sp_check_output(args):
-    proc = sp.Popen(args, stdout=sp.PIPE)
-    out, _ = proc.communicate()
+    proc = sp.Popen(args, stdout=sp.PIPE, stderr=sp.PIPE)
+    out, err = proc.communicate()
     if proc.returncode != 0:
-        raise sp.CalledProcessError(proc.returncode, args, out)
+        msg = "{args} failed.\nStdout:\n{out}\nStderr:\n{err}".format(
+                args=args,
+                out=out,
+                err=err
+            )
+        raise AssertionError(msg)
+    if len(err) > 0:
+        print("STDERR: {0}".format(err))
     return out
 
 def _find_file(name, dir):
@@ -26,6 +37,23 @@ def find_program(name):
         if path is not None:
             return path
     raise Exception("Cannot find program {0}".format(name))
+    
+def using_systemd():
+    try:
+        find_program('systemctl')
+    except:
+        return False
+    else:
+        return True
+
+def get_jobbermaster_logs():
+    if using_systemd():
+        return sp_check_output(['journalctl', '-u', 'jobber'])
+    else:
+        args = ['tail', '-n', '20', '/var/log/messages']
+        lines = sp_check_output(args).split('\n')
+        lines = [l for l in lines if 'jobbermaster' in l]
+        return '/n'.join(lines)
 
 class testlib(object):
     ROBOT_LIBRARY_VERSION = 1.0
@@ -52,21 +80,67 @@ class testlib(object):
         return path
     
     def restart_service(self):
-        # look for systemctl
+        # restart jobber service
         try:
-            systemctl = find_program('systemctl')
-        except:
-            # look for service
-            try:
-                service = find_program('service')
-            except:
-                raise Exception("Don't know how to manage services")
+            if using_systemd():
+                sp_check_output(['systemctl', 'restart', 'jobber'])
             else:
-                # use service
-                sp_check_output([systemctl, 'jobber', 'restart'])
+                sp_check_output(['service', 'jobber', 'restart'])
+        except Exception as e:
+            self.print_debug_info()
+            raise e
+            
+        # wait for it to be ready
+        started = False
+        stop_time = time.time() + 10
+        while time.time() < stop_time and not started:
+            args = [self._jobber_path, 'list']
+            proc = sp.Popen(args, stderr=sp.PIPE)
+            _, err = proc.communicate()
+            if proc.returncode == 0:
+                started = True
+            else:
+                time.sleep(1)
+        if not started:
+            msg = "Failed to start jobber service!"
+            msg += " ('jobber list' returned '{0}')".format(err)
+            raise AssertionError(msg)
+    
+    def print_debug_info(self):
+        log = ''
+    
+        # get service status
+        log += "Jobber service status:\n"
+        if using_systemd():
+            args = ['systemctl', 'status', 'jobber']
         else:
-            # use systemctl
-            sp_check_output([systemctl, 'restart', 'jobber'])
+            args = ['service', 'jobber', 'status']
+        try:
+            log += sp_check_output(args)
+        except:
+            log += "[N/A]"
+            
+        # get syslog msgs
+        log += "\n\njobbermaster logs:\n"
+        try:
+            log += get_jobbermaster_logs()
+        except:
+            log += "[N/A]"
+        
+        # get jobberrunner logs
+        log_files = [
+            _RUNNER_LOG_FILE_FOR_ROOT, 
+            _RUNNER_LOG_FILE_FOR_NORMUSER,
+        ]
+        for lf in log_files:
+            log += "\n\n{0}:\n".format(lf)
+            try:
+                with open(lf) as f:
+                    log += lf.read()
+            except:
+                log += "[N/A]"
+        
+        print(log)
     
     def make_jobfile(self, job_name, cmd, time="*", notify_prog=None):
         jobs_sect = """[jobs]
@@ -120,31 +194,22 @@ notifyProgram: {notify_prog}
 
     def rm_jobfiles(self):
         # rm jobfile
-        try:
+        if os.path.exists(self._root_jobfile_path):
             os.unlink(self._root_jobfile_path)
-        except OSError: pass
-        try:
+        if os.path.exists(self._normuser_jobfile_path):
             os.unlink(self._normuser_jobfile_path)
-        except OSError: pass
-        
-        # reload it
-        sp.check_call([self._jobber_path, 'reload', '-a'], \
-                      stdout=open(os.devnull, 'w'))
     
     def jobber_log(self):
         return sp_check_output([self._jobber_path, 'log'])
     
     def pause_job(self, job):
-        sp.check_call([self._jobber_path, 'pause', job], \
-                      stdout=open(os.devnull, 'w'))
+        sp_check_output([self._jobber_path, 'pause', job])
     
     def resume_job(self, job):
-        sp.check_call([self._jobber_path, 'resume', job], \
-                      stdout=open(os.devnull, 'w'))
+        sp_check_output([self._jobber_path, 'resume', job])
     
     def test_job(self, job):
-        sp.check_call([self._jobber_path, 'test', job], \
-                      stdout=open(os.devnull, 'w'))
+        sp_check_output([self._jobber_path, 'test', job])
     
     def chmod(self, path, mode):
         os.chmod(path, int(mode, base=8))
@@ -176,6 +241,7 @@ notifyProgram: {notify_prog}
         args = [self._jobber_path, 'list']
         if all_users:
             args.append('-a')
+        print("Cmd: {0}".format(args))
         output = sp_check_output(args).strip()
         print(output)
         
@@ -187,6 +253,7 @@ notifyProgram: {notify_prog}
         # do 'jobber list'
         output = sp_check_output(['sudo', '-u', _NORMUSER, \
                                   self._jobber_path, 'list']).strip()
+        print(output)
         
         # check output
         self._check_jobber_list_output(output, job_names)
@@ -202,3 +269,35 @@ notifyProgram: {notify_prog}
             else:
                 raise AssertionError("{msg}: {base_msg}".\
                                     format(msg=msg, base_msg=base_msg))
+    
+    def jobbermaster_has_not_crashed(self):
+        try:
+            logs = get_jobbermaster_logs()
+        except:
+            pass
+        else:
+            if "panic" in logs:
+                print(logs)
+                raise AssertionError("jobbermaster crashed")
+    
+    def jobberrunner_for_root_has_not_crashed(self):
+        try:
+            with open(_RUNNER_LOG_FILE_FOR_ROOT) as f:
+                logs = f.read()
+        except:
+            pass
+        else:
+            if "panic" in logs:
+                print(logs)
+                raise AssertionError("jobberrunner for root crashed")
+    
+    def jobberrunner_for_normuser_has_not_crashed(self):
+        try:
+            with open(_RUNNER_LOG_FILE_FOR_NORMUSER) as f:
+                logs = f.read()
+        except:
+            pass
+        else:
+            if "panic" in logs:
+                print(logs)
+                raise AssertionError("jobberrunner for normuser crashed")
